% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_figure.R
\name{line_figure}
\alias{line_figure}
\title{line_figure}
\usage{
line_figure(
  data_table_list,
  merge_variable,
  value_unit,
  title_name,
  character_list = NULL,
  x_label = "Year",
  lower_limit = 0,
  upper_limit = NA,
  return_static = TRUE,
  source_citation = NULL,
  modifications = NULL,
  subtitle_description = NULL
)
}
\arguments{
\item{subtitle_description}{}
}
\value{
figure
}
\description{
line_figure
}
