% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart_figure_p.R
\name{pie_chart_figure_p}
\alias{pie_chart_figure_p}
\title{pie_chart_figure_p}
\usage{
pie_chart_figure_p(
  data_table_list,
  merge_variable = NULL,
  title_name = NULL,
  character_list = NULL,
  legend_shown = FALSE,
  source_citation = NULL
)
}
\arguments{
\item{source_citation}{}
}
\value{
figure
}
\description{
pie_chart_figure_p
}
