% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_ring_donut_figure_p.R
\name{single_ring_donut_figure_p}
\alias{single_ring_donut_figure_p}
\title{single_ring_donut_figure_p}
\usage{
single_ring_donut_figure_p(
  data_table,
  description_of_goal,
  top_description,
  bottom_description,
  hover_info,
  colors_list,
  character_list = NULL,
  source_citation = NULL
)
}
\arguments{
\item{source_citation}{}
}
\value{
figure
}
\description{
single_ring_donut_figure_p
}
